Remove DiscoverByGenreView
- this was a FrameLayout with a RecyclerView inside (could have just
been a RecyclerView)
- another view bites the dust!
