--- mobile/src/main/res/layout/merge_discover_by_genre.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<merge xmlns:android="http://schemas.android.com/apk/res/android"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent">
-
-  <android.support.v7.widget.RecyclerView
-    android:id="@+id/discover_by_genre_list"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_marginLeft="@dimen/discover_by_genre_list_margin_offset"
-    android:layout_marginTop="@dimen/discover_by_genre_list_margin_offset"
-    android:layout_marginRight="@dimen/discover_by_genre_list_margin_offset"
-    android:layout_marginBottom="@dimen/discover_by_genre_list_margin_offset" />
-
-</merge>
