+++ mobile/src/main/res/layout/view_discover_shows_in_genre_item.xml
